checking concurrency on CI #regressionTest
