+++ .github/workflows/qa-regression-test-edge.yaml
@@ -65,7 +65,7 @@ jobs:
           time: 15s
 
       - name: UI regression tests
-        run: mvn test -Dgroups="exampleTag" -Dbrowser="edge"
+        run: mvn clean test -T 4 -Dgroups="TokenBalanceUI" -Dbrowser="edge" -Dheadless="enable" -Dnetwork="disable" "-Djunit.jupiter.execution.parallel.enabled=true"
         working-directory: ./quality-assurance/qa-ui
 
       - name: Upload allure-results
