+++ quality-assurance/qa-ui/src/test/java/pl/kargolek/extension/javascript/JavascriptDriverExtension.java
@@ -1,32 +1,16 @@
 package pl.kargolek.extension.javascript;
 
-import org.junit.jupiter.api.extension.*;
-import pl.kargolek.extension.util.AnnotationResolver;
+import org.junit.jupiter.api.extension.ExtensionContext;
+import org.junit.jupiter.api.extension.ParameterContext;
+import org.junit.jupiter.api.extension.ParameterResolutionException;
+import org.junit.jupiter.api.extension.ParameterResolver;
 import pl.kargolek.util.JavascriptDriver;
-import pl.kargolek.util.WebDriverResolver;
+import pl.kargolek.util.WebDriverFactory;
 
 /**
  * @author Karol Kuta-Orlowicz
  */
-public class JavascriptDriverExtension implements ParameterResolver, BeforeAllCallback, BeforeEachCallback {
-
-    private JavascriptDriver javascriptDriver;
-    private final WebDriverResolver webDriverResolver = new WebDriverResolver();
-    private final AnnotationResolver annotationResolver = new AnnotationResolver();
-
-    @Override
-    public void beforeAll(ExtensionContext extensionContext) {
-        if (annotationResolver.getSeleniumWebDriverAnnotation(extensionContext).isBeforeAll()) {
-            initDriver(extensionContext);
-        }
-    }
-
-    @Override
-    public void beforeEach(ExtensionContext extensionContext) {
-        if (!annotationResolver.getSeleniumWebDriverAnnotation(extensionContext).isBeforeAll()) {
-            initDriver(extensionContext);
-        }
-    }
+public class JavascriptDriverExtension implements ParameterResolver {
 
     @Override
     public boolean supportsParameter(ParameterContext parameterContext, ExtensionContext extensionContext) throws ParameterResolutionException {
@@ -35,14 +19,9 @@ public class JavascriptDriverExtension implements ParameterResolver, BeforeAllCa
 
     @Override
     public Object resolveParameter(ParameterContext parameterContext, ExtensionContext extensionContext) throws ParameterResolutionException {
-        return this.javascriptDriver;
-    }
-
-    private void initDriver(ExtensionContext context) {
-        var driver = this.webDriverResolver.getStoredWebDriver(context);
-        this.javascriptDriver = JavascriptDriver
-                .builder(driver)
+        return JavascriptDriver.builder(WebDriverFactory.getRemoteWebDriverInstance())
                 .createJavascriptExecutor()
                 .build();
     }
+
 }
