+++ quality-assurance/qa-ui/src/test/java/pl/kargolek/extension/logs/BrowserLogsExtension.java
@@ -4,10 +4,10 @@ import org.junit.jupiter.api.extension.AfterTestExecutionCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.openqa.selenium.logging.LogEntry;
 import org.openqa.selenium.logging.LogType;
-import pl.kargolek.util.BrowserType;
 import pl.kargolek.util.ReportAttachment;
 import pl.kargolek.util.TestProperty;
-import pl.kargolek.util.WebDriverResolver;
+import pl.kargolek.util.WebDriverFactory;
+import pl.kargolek.util.constant.BrowserType;
 
 import java.nio.charset.StandardCharsets;
 import java.time.Instant;
@@ -19,16 +19,14 @@ import java.util.List;
  * @author Karol Kuta-Orlowicz
  */
 public class BrowserLogsExtension implements AfterTestExecutionCallback {
-    private final WebDriverResolver webDriverResolver = new WebDriverResolver();
+
     private final ReportAttachment reportAttachment = new ReportAttachment();
     private final BrowserType browserType = TestProperty.getInstance().getBrowserType();
 
     @Override
     public void afterTestExecution(ExtensionContext extensionContext) {
         if (isDriverSupportBrowserLog()) {
-            var driver = webDriverResolver
-                    .getStoredWebDriver(extensionContext);
-
+            var driver = WebDriverFactory.getRemoteWebDriverInstance();
             var consoleLogs = driver
                     .manage()
                     .logs()
