+++ quality-assurance/qa-ui/src/test/java/pl/kargolek/tests/InputWalletsUIAsyncTest.java
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.parallel.Execution;
+import org.junit.jupiter.api.parallel.ExecutionMode;
 import org.openqa.selenium.WebDriver;
 import pl.kargolek.data.script.TestDataSql;
 import pl.kargolek.extension.BaseTestConfig;
@@ -24,7 +26,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 @Feature("Input wallet UI")
 @BaseTestConfig
 @SoftAssertion
-public class InputWalletsUITest {
+public class InputWalletsUIAsyncTest {
 
     private String appBaseURL;
     private InitPages pages;
@@ -47,6 +49,7 @@ public class InputWalletsUITest {
     @Severity(SeverityLevel.NORMAL)
     @Story("As a normal user, I want to see app title")
     @Description("When user open wallet's provider page, app title should be display")
+    @Execution(ExecutionMode.CONCURRENT)
     void whenOpenWalletPage_thenTitleShouldBeDisplay(TestProperty property) {
         var appTitleText = this.pages.getHomePage()
                 .open(this.appBaseURL)
@@ -59,6 +62,7 @@ public class InputWalletsUITest {
     @Severity(SeverityLevel.NORMAL)
     @Story("As a normal user, I want to see description which tokens are handle")
     @Description("Handled tokens names should be provided below input wallet addresses field")
+    @Execution(ExecutionMode.CONCURRENT)
     void whenOpenWalletPage_thenTokenHandleDescriptionShouldDisplayed(SoftAssertions softAssertions) {
         var tokenDescriptionText = this.pages.getHomePage()
                 .open(this.appBaseURL)
@@ -75,6 +79,7 @@ public class InputWalletsUITest {
     @Severity(SeverityLevel.NORMAL)
     @Story("As a normal user, I can see placeholder in input field")
     @Description("User can see placeholder in input field that inform what data should be typed")
+    @Execution(ExecutionMode.CONCURRENT)
     void whenOpenWalletPage_thenInputPlaceholderShouldBeVisible() {
         var placeholder = this.pages.getHomePage()
                 .open(this.appBaseURL)
@@ -88,6 +93,7 @@ public class InputWalletsUITest {
     @Story("As a normal user, when I type ETH wallet address and press send button " +
             "I want to open balance page")
     @Description("When user hit send button with valid ETH wallet address, balance page should be open")
+    @Execution(ExecutionMode.CONCURRENT)
     void whenWalletAddressCorrectAndPressSend_thenBalancePageShouldBeOpen() {
         var isBalancePageOpened = this.pages.getHomePage()
                 .open(this.appBaseURL)
@@ -104,6 +110,7 @@ public class InputWalletsUITest {
     @Story("As a normal user, when I type ETH wallet address and press enter key " +
             "I want to open balance page")
     @Description("When user hit send key with valid ETH wallet address, balance page should be open")
+    @Execution(ExecutionMode.CONCURRENT)
     void whenWalletAddressCorrectAndPressEnterKey_thenBalancePageShouldBeOpen() {
         var isBalancePageOpened = this.pages.getHomePage()
                 .open(this.appBaseURL)
@@ -120,6 +127,7 @@ public class InputWalletsUITest {
     @Story("As a normal user, when I type ETH wallet address refresh page or back from other page " +
             "then wallet address should be maintained in the input wallets")
     @Description("Wallet address should be saved in browser session storage")
+    @Execution(ExecutionMode.CONCURRENT)
     void whenSendWalletAddress_thenShouldBeSavedInTheSessionStorage(WebDriver driver, SoftAssertions softAssertions) {
         var inputWalletPage = this.pages.getInputWalletPage();
         var walletAddress = "0x6cc28D37607024098F4104228E1388953875309B";
@@ -149,6 +157,7 @@ public class InputWalletsUITest {
     @Severity(SeverityLevel.NORMAL)
     @Story("As a normal user, I can use navbar in wallets page")
     @Description("Navbar is enable in the wallets page")
+    @Execution(ExecutionMode.CONCURRENT)
     void whenOpenWalletPage_thenNavbarShouldBeDisplay() {
         this.pages.getHomePage()
                 .open(this.appBaseURL);
@@ -162,6 +171,7 @@ public class InputWalletsUITest {
     @Severity(SeverityLevel.NORMAL)
     @Story("As a normal user when open wallet input page, I can use price banner")
     @Description("Price banner component should be attached to /wallet page")
+    @Execution(ExecutionMode.CONCURRENT)
     void whenOpenWalletPage_thenPriceBannerShouldBeAvailable() {
         this.pages.getHomePage()
                 .open(this.appBaseURL);
