+++ quality-assurance/qa-ui/src/test/java/pl/kargolek/util/CustomParallelStrategy.java
@@ -1,0 +1,41 @@
+package pl.kargolek.util;
+
+import org.junit.platform.engine.ConfigurationParameters;
+import org.junit.platform.engine.support.hierarchical.ParallelExecutionConfiguration;
+import org.junit.platform.engine.support.hierarchical.ParallelExecutionConfigurationStrategy;
+
+/**
+ * @author Karol Kuta-Orlowicz
+ */
+public class CustomParallelStrategy implements ParallelExecutionConfiguration, ParallelExecutionConfigurationStrategy {
+
+    @Override
+    public int getParallelism() {
+        return 2;
+    }
+
+    @Override
+    public int getMinimumRunnable() {
+        return 0;
+    }
+
+    @Override
+    public int getMaxPoolSize() {
+        return 2;
+    }
+
+    @Override
+    public int getCorePoolSize() {
+        return 2;
+    }
+
+    @Override
+    public int getKeepAliveSeconds() {
+        return 30;
+    }
+
+    @Override
+    public ParallelExecutionConfiguration createConfiguration(final ConfigurationParameters configurationParameters) {
+        return this;
+    }
+}
