+++ quality-assurance/qa-ui/src/test/java/pl/kargolek/util/TestProperty.java
@@ -1,5 +1,12 @@
 package pl.kargolek.util;
 
+import pl.kargolek.util.constant.BrowserType;
+import pl.kargolek.util.constant.HeadlessMode;
+import pl.kargolek.util.constant.NetworkReporter;
+import pl.kargolek.util.constant.ParallelTest;
+
+import java.util.Optional;
+
 public class TestProperty {
     private static final TestProperty instance = new TestProperty();
     private static final PropertiesLoader PROPERTIES_LOADER = new PropertiesLoader(PathResolver.TEST_RESOURCES_PATH,
@@ -36,16 +43,32 @@ public class TestProperty {
     }
 
     public BrowserType getBrowserType(){
-        var browser = System.getProperty("browser");
-        if (browser == null){
-            browser = "chrome";
-        }
+        var browser = Optional.ofNullable(System.getProperty("browser"))
+                .orElse("chrome");
         return switch (browser) {
             case "firefox" -> BrowserType.FIREFOX;
             case "edge" -> BrowserType.EDGE;
             case "mobile-chrome" -> BrowserType.MOBILE_CHROME;
             default -> BrowserType.CHROME;
         };
+    }
+
+    public HeadlessMode getHeadlessMode(){
+        var headlessMode = Optional.ofNullable(System.getProperty("headless"))
+                .orElse("disable");
+        return headlessMode.equals("enable") ? HeadlessMode.ENABLE : HeadlessMode.DISABLE;
+    }
+
+    public ParallelTest getParallelism(){
+        var parallel = Optional.ofNullable(System.getProperty("junit.jupiter.execution.parallel.enabled"))
+                .orElse("false");
+        return parallel.equals("true") ? ParallelTest.ENABLE : ParallelTest.DISABLE;
+    }
+
+    public NetworkReporter getNetworkReporter(){
+        var network = Optional.ofNullable(System.getProperty("network"))
+                .orElse("enable");
+        return network.equals("enable") ? NetworkReporter.ENABLE : NetworkReporter.DISABLE;
     }
