+++ quality-assurance/qa-ui/src/test/java/pl/kargolek/util/constant/HeadlessMode.java
@@ -1,0 +1,8 @@
+package pl.kargolek.util.constant;
+
+/**
+ * @author Karol Kuta-Orlowicz
+ */
+public enum HeadlessMode {
+    ENABLE,DISABLE
+}
