+++ quality-assurance/qa-ui/src/test/java/pl/kargolek/util/constant/NetworkReporter.java
@@ -1,0 +1,8 @@
+package pl.kargolek.util.constant;
+
+/**
+ * @author Karol Kuta-Orlowicz
+ */
+public enum NetworkReporter {
+    ENABLE, DISABLE;
+}
