+++ quality-assurance/qa-ui/src/test/resources/junit-platform.properties
@@ -1,0 +1,5 @@
+junit.jupiter.execution.parallel.enabled=false
+junit.jupiter.execution.parallel.mode.default=concurrent
+junit.jupiter.execution.parallel.mode.classes.default=same_thread
+junit.jupiter.execution.parallel.config.strategy = custom
+junit.jupiter.execution.parallel.config.custom.class = pl.kargolek.util.CustomParallelStrategy
